23/07 update contract

+++ src/main/java/com/gcl/crm/repository/CustomerProcessRepository.java
@@ -4,7 +4,7 @@ package com.gcl.crm.repository;
 import com.gcl.crm.entity.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerProcessRepository  extends JpaRepository<Customer,Long> {
+public interface CustomerProcessRepository  extends JpaRepository<Customer,Integer> {
 
 
 }
